Remove unnecessary inheritance for HttpContent
